Move get_checksum() to utils.py module
